+++ apps/docs/src/consts.ts
@@ -1,0 +1,1 @@
+export const CONTACT_URL = 'https://www.synergycodes.com/contact';