+++ packages/ng-diagram/projects/ng-diagram/src/core/src/input-events/handlers/wheel-panning/wheel-panning.event.ts
@@ -1,0 +1,7 @@
+import { BaseInputEvent } from '../../input-events.interface';
+
+export interface WheelPanningEvent extends BaseInputEvent {
+  name: 'wheelPanning';
+  deltaX: number;
+  deltaY: number;
+}